+++ src/app/panel/tarifas/[feeId]/page.tsx
@@ -1,8 +1,16 @@
 import { Title } from "~/components/title";
 import { api } from "~/trpc/server";
 import FeePage from "./fee-page";
+import { tienePermiso } from "~/lib/permisos";
+import { redirect } from "next/navigation";
 
 export default async function Channel(props: { params: { feeId: string } }) {
+  const { perms } = await api.user.self.query();
+  if (!tienePermiso(perms, "panel:locales")) {
+    redirect("/accessdenied");
+    return <></>;
+  }
+
   const fee = await api.fee.getByIdProt.query({
     id: props.params.feeId,
   });
