+++ src/app/panel/usuarios/page.tsx
@@ -1,11 +1,19 @@
 // import { auth, clerkClient } from "@clerk/nextjs/server";
+import { redirect } from "next/navigation";
 import { List, ListTile } from "~/components/list";
 import { Title } from "~/components/title";
 import { Badge } from "~/components/ui/badge";
 import { Roles } from "~/lib/globals";
+import { PERMISO_ADMIN, tienePermiso } from "~/lib/permisos";
 import { api } from "~/trpc/server";
 
 export default async function Home() {
+  const { perms } = await api.user.self.query();
+  if (!tienePermiso(perms, PERMISO_ADMIN)) {
+    redirect("/accessdenied");
+    return <></>;
+  }
+
   // const response = await clerkClient.users.getUserList();
   await api.test.migrateToEntities.mutate();
   const users = await api.user.listBasic.query();
